pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "matplotlib>=3.10.1",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "requests>=2.32.3",
    "trafilatura>=2.0.0",
    "flask-wtf>=1.2.2",
    "sqlalchemy>=2.0.40",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "notion-client>=2.3.0",
    "pytz>=2025.2",
    "schedule>=1.2.2",
    "flask-dance>=7.1.0",
    "oauthlib>=3.3.1",
    "pyjwt>=2.10.1",
    "websocket-client>=1.8.0",
    "websocket>=0.2.1",
    "sendgrid>=6.12.4",
    "slack-sdk>=3.35.0",
    "scikit-learn>=1.7.0",
    "wkhtmltopdf>=0.2",
    "reportlab>=4.4.2",
    "python-pptx>=1.0.2",
    "jinja2>=3.1.6",
    "plotly>=6.2.0",
    "statsmodels>=0.14.4",
    "gitpython>=3.1.44",
    "psutil>=7.0.0",
    "pytest>=8.4.1",
    "aiohttp>=3.12.15",
    "werkzeug>=3.1.3",
    "openpyxl>=3.1.5",
    "xlsxwriter>=3.2.5",
    "google-api-python-client>=2.179.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "web3>=7.13.0",
    "eth-account>=0.13.7",
    "eth-abi>=5.2.0",
    "eth-typing>=5.2.1",
    "cryptography>=46.0.1",
    "user-agents>=2.2.0",
    "qrcode>=8.2",
    "pillow>=11.2.1",
    "bitcoinlib>=0.7.3",
    "redis>=3.5.3",
    "redis-py-cluster>=2.1.3",
    "markdown-pdf>=1.10",
    "weasyprint>=66.0",
    "markdown>=3.9",
    "rq>=2.6.0",
    "xgboost>=3.0.5",
    "pulp>=3.3.0",
    "flask-caching>=2.3.1",
    "apscheduler>=3.11.1",
    "python-dateutil>=2.9.0.post0",
    "pynacl>=1.6.1",
    "flake8>=7.3.0",
]
